requirements.txt
asgiref==3.9.1
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
joblib==1.5.2
nltk==3.9.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
regex==2025.7.34
sqlparse==0.5.3
textblob==0.19.0
tqdm==4.67.1
tzdata==2025.2
uvicorn==0.35.0
whitenoise==6.10.0
